refactor(runbotics-orchestrator-ui): :recycle: Fix after CR
